Replace svg2img with sharp to avoid recompiling native canvas module
node-canvas module is a dependency of svg2img. svg2img was being used
just be command-line maestro for generating summary timeline images in
docx files. Electron uses a different version of node than is likely
installed at the system level, and as such, when the maestro-cli was
used to generate a docx it required one compilation of node-canvas for
the system-level Node and a different for Electron's Node. There is
probably no good way to satisfy both while Maestro remains a single
repository. Furthermore, rebuilding node-canvas on Windows is a pain.
Swapping out for 'sharp' simplifies things.